Added the insertion of all values to mySQL.
Script is fully functional (but will continue to be updated and improved) and will be used only for the first run. This will run through all IDs from 0-50k to find out which IDs correlate to tradable items (as well as sending their info to our database). After that, the primary_scraper will take over, only iterating through the list of tradable items, which will substantially reduce the duration of each run.